# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ArtFi API
  version: 0.4.0
  description: Stage 4 Base Sepolia preview contract. Wallet intents never hold private keys; authenticated indexer writes are durable and replay-safe.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/uploads/intents:
    post:
      operationId: createUploadIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadIntentRequest"
      responses:
        "201":
          description: Short-lived digest-bound API upload intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadIntent"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/uploads/{uploadID}:
    put:
      operationId: uploadObject
      parameters:
        - name: uploadID
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{32}$" }
        - name: Content-SHA256
          in: header
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{64}$" }
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "204": { description: Object digest verified and persisted }
        "400":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/rwa/intents:
    post:
      operationId: createMintIntent
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MintIntentRequest"
      responses:
        "201":
          description: New immutable metadata and contract-call intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "200":
          description: Idempotent replay of an existing intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/rwa/intents/{intentID}:
    get:
      operationId: getMintIntent
      parameters:
        - $ref: "#/components/parameters/IntentID"
      responses:
        "200":
          description: Mint transaction lifecycle snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    IntentID:
      name: intentID
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9a-f]{32}$" }
  responses:
    Problem:
      description: RFC 9457-style request failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    UploadIntentRequest:
      type: object
      additionalProperties: false
      required: [fileName, contentType, sha256, size]
      properties:
        fileName: { type: string, minLength: 1, maxLength: 120 }
        contentType: { type: string, enum: [image/png, image/jpeg, image/webp] }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        size: { type: integer, minimum: 1, maximum: 10485760 }
    UploadIntent:
      type: object
      required:
        [uploadId, uploadUrl, method, expiresAt, maxBytes, requiredHeaders]
      properties:
        uploadId: { type: string, pattern: "^[0-9a-f]{32}$" }
        uploadUrl: { type: string }
        method: { type: string, const: PUT }
        expiresAt: { type: string, format: date-time }
        maxBytes: { type: integer, const: 10485760 }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
    MintIntentRequest:
      type: object
      additionalProperties: false
      required:
        [uploadId, recipient, name, artist, year, medium, location, description]
      properties:
        uploadId: { type: string, pattern: "^[0-9a-f]{32}$" }
        recipient: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        name: { type: string, minLength: 2, maxLength: 120 }
        artist: { type: string, minLength: 2, maxLength: 120 }
        year: { type: integer, minimum: 1000 }
        medium: { type: string, minLength: 2, maxLength: 160 }
        location: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, minLength: 20, maxLength: 2000 }
    MintIntent:
      type: object
      required:
        [
          intentId,
          requestId,
          recipient,
          registryAddress,
          chainId,
          metadataUri,
          metadataSha256,
          contractFunction,
          contractArguments,
          status,
          createdAt,
        ]
      properties:
        intentId: { type: string, pattern: "^[0-9a-f]{32}$" }
        requestId: { type: string, pattern: "^0x[0-9a-f]{64}$" }
        recipient: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        registryAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        chainId: { type: integer, const: 84532 }
        metadataUri: { type: string, format: uri, pattern: "^https://" }
        metadataSha256: { type: string, pattern: "^0x[0-9a-f]{64}$" }
        contractFunction:
          { type: string, const: "createAsset(bytes32,address,string,bytes32)" }
        contractArguments:
          type: array
          minItems: 4
          maxItems: 4
          items: { type: string }
        status: { type: string, enum: [prepared, submitted, confirmed, failed] }
        transactionHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        createdAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }